remove recursive version of repo fetcher; move promise based repo fetcher to separate file
